refactor appointment date provider and filter strategies (#29495)
