+++ packages/devextreme/js/__internal/scheduler/appointments/data_provider/m_appointment_data_provider.ts
@@ -3,11 +3,18 @@ import { combineRemoteFilter } from '@ts/scheduler/r1/filterting/index';
 import type { AppointmentDataAccessor } from '@ts/scheduler/utils';
 
 import { AppointmentDataSource } from './m_appointment_data_source';
-import { AppointmentFilterBaseStrategy, AppointmentFilterVirtualStrategy } from './m_appointment_filter';
+import { AppointmentFilterBaseStrategy } from './m_appointment_filter';
+import { AppointmentFilterVirtualStrategy } from './m_appointment_filter_virtual';
 
-const FilterStrategies = {
-  virtual: 'virtual',
-  standard: 'standard',
+type FilterStrategy = (AppointmentFilterBaseStrategy | AppointmentFilterVirtualStrategy) & {
+  constructor: Function & {
+    strategyName: string;
+  };
+};
+
+const FilterStrategyMap = {
+  [AppointmentFilterVirtualStrategy.strategyName]: AppointmentFilterVirtualStrategy,
+  [AppointmentFilterBaseStrategy.strategyName]: AppointmentFilterBaseStrategy,
 };
 
 export class AppointmentDataProvider {
@@ -21,7 +28,7 @@ export class AppointmentDataProvider {
 
   appointmentDataSource: AppointmentDataSource;
 
-  filterStrategy: any;
+  filterStrategy!: FilterStrategy;
 
   constructor(options) {
     this.options = options;
@@ -30,7 +37,6 @@ export class AppointmentDataProvider {
     this.timeZoneCalculator = this.options.timeZoneCalculator;
 
     this.appointmentDataSource = new AppointmentDataSource(this.dataSource);
-
     this.initFilterStrategy();
   }
 
@@ -42,21 +48,24 @@ export class AppointmentDataProvider {
     return !!this.dataSource;
   }
 
-  get filterStrategyName() {
+  get filterStrategyName(): string {
     return this.options.getIsVirtualScrolling()
-      ? FilterStrategies.virtual
-      : FilterStrategies.standard;
+      ? AppointmentFilterVirtualStrategy.strategyName
+      : AppointmentFilterBaseStrategy.strategyName;
   }
 
-  getFilterStrategy() {
-    if (!this.filterStrategy || this.filterStrategy.strategyName !== this.filterStrategyName) {
+  getFilterStrategy(): AppointmentFilterBaseStrategy | AppointmentFilterVirtualStrategy {
+    if (
+      !this.filterStrategy
+      || this.filterStrategy.constructor.strategyName !== this.filterStrategyName
+    ) {
       this.initFilterStrategy();
     }
 
     return this.filterStrategy;
   }
 
-  initFilterStrategy() {
+  initFilterStrategy(): void {
     const filterOptions = {
       resources: this.options.resources,
       dataAccessors: this.dataAccessors,
@@ -75,19 +84,18 @@ export class AppointmentDataProvider {
       viewDataProvider: this.options.getViewDataProvider,
       allDayPanelMode: this.options.allDayPanelMode,
     };
+    const strategy = new FilterStrategyMap[this.filterStrategyName](filterOptions);
 
-    this.filterStrategy = this.filterStrategyName === FilterStrategies.virtual
-      ? new AppointmentFilterVirtualStrategy(filterOptions)
-      : new AppointmentFilterBaseStrategy(filterOptions);
+    this.filterStrategy = strategy as FilterStrategy;
   }
 
-  setDataSource(dataSource) {
+  setDataSource(dataSource): void {
     this.dataSource = dataSource;
     this.initFilterStrategy();
     this.appointmentDataSource.setDataSource(this.dataSource);
   }
 
-  updateDataAccessors(dataAccessors) {
+  updateDataAccessors(dataAccessors): void {
     this.dataAccessors = dataAccessors;
     this.initFilterStrategy();
   }
@@ -122,10 +130,6 @@ export class AppointmentDataProvider {
 
   filterLoadedAppointments(filterOption, preparedItems) {
     return this.getFilterStrategy().filterLoadedAppointments(filterOption, preparedItems);
-  }
-
-  calculateAppointmentEndDate(isAllDay, startDate) {
-    return this.getFilterStrategy().calculateAppointmentEndDate(isAllDay, startDate);
   }
 
   // Appointment data source mappings
