+++ packages/devextreme/js/__internal/scheduler/m_scheduler.ts
@@ -37,11 +37,14 @@ import errors from '@js/ui/widget/ui.errors';
 import Widget from '@js/ui/widget/ui.widget';
 import { dateUtilsTs } from '@ts/core/utils/date';
 import { createTimeZoneCalculator } from '@ts/scheduler/r1/timezone_calculator/index';
+import type { AppointmentDataItem } from '@ts/scheduler/r1/types';
 import {
   excludeFromRecurrence,
-  getAppointmentTakesAllDay,
   getPreparedDataItems,
-  isDateAndTimeView, isTimelineView, viewsUtils,
+  isAppointmentTakesAllDay,
+  isDateAndTimeView,
+  isTimelineView,
+  viewsUtils,
 } from '@ts/scheduler/r1/utils/index';
 import type { IFieldExpr } from '@ts/scheduler/utils/index';
 import { macroTaskArray } from '@ts/scheduler/utils/index';
@@ -160,7 +163,7 @@ const RECURRENCE_EDITING_MODE = {
 class Scheduler extends Widget<any> {
   _filteredItems!: any[];
 
-  _preparedItems!: any[];
+  _preparedItems!: AppointmentDataItem[];
 
   _timeZoneCalculator!: any;
 
@@ -172,7 +175,7 @@ class Scheduler extends Widget<any> {
 
   _appointments: any;
 
-  appointmentDataProvider: any;
+  appointmentDataProvider!: AppointmentDataProvider;
 
   _dataSource: any;
 
@@ -404,14 +407,14 @@ class Scheduler extends Widget<any> {
     this._filteredItems = value;
   }
 
-  get preparedItems() {
+  get preparedItems(): AppointmentDataItem[] {
     if (!this._preparedItems) {
       this._preparedItems = [];
     }
     return this._preparedItems;
   }
 
-  set preparedItems(value) {
+  set preparedItems(value: AppointmentDataItem[]) {
     this._preparedItems = value;
   }
 
@@ -1196,11 +1199,10 @@ class Scheduler extends Widget<any> {
 
     workspace.option('allDayExpanded', this._isAllDayExpanded());
 
-    let viewModel = [];
     // @ts-expect-error
-    if (this._isVisible()) {
-      viewModel = this._getAppointmentsToRepaint();
-    }
+    const viewModel = this._isVisible()
+      ? this._getAppointmentsToRepaint()
+      : [];
 
     this._appointments.option('items', viewModel);
     this.appointmentDataProvider.cleanState();
@@ -1726,7 +1728,6 @@ class Scheduler extends Widget<any> {
       indicatorTime: this.option('indicatorTime'),
       indicatorUpdateInterval: this.option('indicatorUpdateInterval'),
       shadeUntilCurrentTime: this.option('shadeUntilCurrentTime'),
-      allDayExpanded: this._appointments.option('items'),
       crossScrollingEnabled,
       dataCellTemplate: this.option('dataCellTemplate'),
       timeCellTemplate: this.option('timeCellTemplate'),
@@ -2260,7 +2261,7 @@ class Scheduler extends Widget<any> {
       this.timeZoneCalculator,
     );
 
-    return getAppointmentTakesAllDay(
+    return isAppointmentTakesAllDay(
       appointment,
       this._getCurrentViewOption('allDayPanelMode'),
     );
